Codechange: Increase scrollbar's limit to 2**32 and make it properly unsigned

+++ src/bridge_gui.cpp
@@ -153,7 +153,7 @@ public:
 		this->bridges->NeedResort();
 		this->SortBridgeList();
 
-		this->vscroll->SetCount((uint)bl->size());
+		this->vscroll->SetCount(bl->size());
 	}
 
 	~BuildBridgeWindow()
