+++ src/town_gui.cpp
@@ -700,7 +700,7 @@ private:
 
 			this->towns.shrink_to_fit();
 			this->towns.RebuildDone();
-			this->vscroll->SetCount((uint)this->towns.size()); // Update scrollbar as well.
+			this->vscroll->SetCount(this->towns.size()); // Update scrollbar as well.
 		}
 		/* Always sort the towns. */
 		this->towns.Sort();
@@ -793,7 +793,7 @@ public:
 				break;
 
 			case WID_TD_LIST: {
-				int n = 0;
+				uint n = 0;
 				int y = r.top + WD_FRAMERECT_TOP;
 				if (this->towns.size() == 0) { // No towns available.
 					DrawString(r.left + WD_FRAMERECT_LEFT, r.right, y, STR_TOWN_DIRECTORY_NONE);
